fix: InsecureTransportError raising
There was an issue with InsecureTransportError being raised while the
request has not fully been initialized by Django/Flask. Then the
authorization server would try to catch the exception and enrich with
request.payload.state, that don't exist because the request is not
initialized.

The fix is to avoid enriching the 'state' parameter for exceptions
raised during the request initialization, that for the moment can only
be InsecureTransportError.